2.0.54 backport

+++ core/src/test/java/com/alibaba/fastjson2/JSONWriterJSONBTest.java
@@ -198,11 +198,11 @@ public class JSONWriterJSONBTest {
             JSONWriterJSONB jsonWriter = (JSONWriterJSONB) JSONWriter.ofJSONB();
             bytes.set(jsonWriter, new byte[0]);
             jsonWriter.writeInt64(new long[1]);
-            jsonWriter.writeInt64(null);
+            jsonWriter.writeInt64((Long) null);
             jsonWriter.writeFloat((Float) null);
             jsonWriter.writeFloat((float[]) null);
             jsonWriter.writeDouble((double[]) null);
-            jsonWriter.writeInt32(null);
+            jsonWriter.writeInt32((Integer) null);
             jsonWriter.writeLocalDateTime(null);
             jsonWriter.writeUUID(null);
             jsonWriter.writeBigInt(null);
